+++ changelog.md
@@ -13,6 +13,7 @@ __[maintenance]__
 * Update certificate information for Eclipse Temurin JRE 11 LTS installer.
 * Update certificate information for Eclipse Temurin JDK 17 LTS installer.
 * Update certificate information for Eclipse Temurin JRE 17 LTS installer.
+* Update certificate information for Eclipse Temurin JDK 21 LTS installer.
 
 ## Version 2026.07.18.0